Add back link to steps
With the former structure I would have needed to change every single
route to make it pass the back link url to the generic render function
from _step.
The new structure kind of reverses the "command chain". Now the generic
render function (Step) in _step is called first and determines which
step specific components to render. Now I only have to pass the back
link once to the render function.