Add handling of Ekona response
We want to handle the Ekona response. Unfortunately, Ekona does not send signed assertions when we set the flag wantAssertionsSigned. Therefore, this can not be checked and the configuration of the Saml component is adapted.

This also adds the storage of the userId in the cookie and a function to retrieve the user by id from the database